requirements.txt
pandas>=2.2,<3.0
numpy>=1.26,<3.0
scikit-learn>=1.5,<2.0
pyarrow>=17.0,<20.0
kagglehub>=0.3,<1.0
mlflow>=2.16,<3.0
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
kafka-python>=2.0,<3.0
pydantic>=2.8,<3.0
pytest>=8.2,<9.0

httpx>=0.27,<1.0
types-requests>=2.32.0,<3.0
